Finish k8s user–friendly storage support
Works like the other providers except for two “insignificant” details:

1. Users need to specify an existing StorageClass and storage capacity
to allocate with a slightly different syntax: `directory = "sc:1:/path"`
would create a PersistentVolumeClaim from the StorageClass `sc` with `1`
GB of capacity, and synchronize `/path` from the user's computer back
and forth. The used StorageClass should support the ReadWriteOnce access
mode for `parallelism = 1` or ReadWriteMany for `parallelism > 1` in
order to be accessed by different pods at the same time. In the first
case, rancher/local-path-provisioner can be used for testing purposes.

2. There is a “feature” (read: bug) pulling files that creates nested
directories. Solving it is trivial, but we need to implement on our own
some of the `kubectl cp` logic.